2mm float

+++ xla/examples/polybench/2mm.cc
@@ -130,7 +130,7 @@ int main(int argc, char** argv)
 
   auto beta_b = buildBufferFromScalar(client, beta);
 
-  auto tmp_a = xla::Array2D<double>(ni, nj);
+  auto tmp_a = xla::Array2D<float>(ni, nj);
   for (int i = 0; i < ni; i++) {
     for (int j = 0; j < nj; j++) {
       tmp_a(i, j) = (*tmp)[i][j];
@@ -138,25 +138,25 @@ int main(int argc, char** argv)
   }
   auto tmp_b = buildBuffer2D(client, tmp_a);
 
-  auto A_a = xla::Array2D<double>(ni, nk);
+  auto A_a = xla::Array2D<float>(ni, nk);
   for (int i = 0; i < ni; i++)
     for (int j = 0; j < nk; j++)
       A_a(i, j) = (*A)[i][j];
   auto A_b = buildBuffer2D(client, A_a);
 
-  auto B_a = xla::Array2D<double>(nk, nj);
+  auto B_a = xla::Array2D<float>(nk, nj);
   for (int i = 0; i < nk; i++)
     for (int j = 0; j < nj; j++)
       B_a(i, j) = (*B)[i][j];
   auto B_b = buildBuffer2D(client, B_a);
 
-  auto C_a = xla::Array2D<double>(nj, nl);
+  auto C_a = xla::Array2D<float>(nj, nl);
   for (int i = 0; i < nj; i++)
     for (int j = 0; j < nl; j++)
       C_a(i, j) = (*C)[i][j];
   auto C_b = buildBuffer2D(client, C_a);
 
-  auto D_a = xla::Array2D<double>(ni, nl);
+  auto D_a = xla::Array2D<float>(ni, nl);
   for (int i = 0; i < ni; i++)
     for (int j = 0; j < nl; j++)
       D_a(i, j) = (*D)[i][j];
@@ -182,7 +182,7 @@ int main(int argc, char** argv)
   /* Store the result data. */
   std::shared_ptr<Literal> result_literal = 
                           result[0][0]->ToLiteralSync().value();
-  auto result_a = result_literal->data<double>();
+  auto result_a = result_literal->data<float>();
   for (int i = 0; i < ni; i++)
     for (int j = 0; j < nl; j++)
       (*D)[i][j] = result_a[i * nl + j];
